refactor: move stuff around
